Build_4
There was no nightly yesterday, however the game is so much better graphically.
Many of the 8-bit sprites have been changed into 16-bit sprites. The map has seen a 400% increase in size, with better grass textures and more detail. There also is a sizeable lake around the middle to the top right of the map, and the house is still under renovation: don't go in because it's still buggy.